Rename to RemoteAirDataRecord and AirData for remote/local class
