+++ app/src/main/java/tw/wesley/uiassignment/data/database/AirDataRecordDao.kt
@@ -1,21 +1,21 @@
-package tw.wesley.uiassignment.data.daos
+package tw.wesley.uiassignment.data.database
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import tw.wesley.uiassignment.data.entities.AirDataRecordEntity
+import tw.wesley.uiassignment.data.local.AirData
 
 @Dao
 interface AirDataRecordDao {
 
     @Query("SELECT * FROM air_data_records")
-    fun getAllRecords(): List<AirDataRecordEntity>
+    fun getAllRecords(): List<AirData>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertAll(records: List<AirDataRecordEntity>)
+    suspend fun insertAll(records: List<AirData>)
 
     @Delete
-    suspend fun delete(record: AirDataRecordEntity)
+    suspend fun delete(record: AirData)
 }
