+++ app/src/main/java/tw/wesley/uiassignment/data/database/AppDatabase.kt
@@ -2,10 +2,9 @@ package tw.wesley.uiassignment.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import tw.wesley.uiassignment.data.daos.AirDataRecordDao
-import tw.wesley.uiassignment.data.entities.AirDataRecordEntity
+import tw.wesley.uiassignment.data.local.AirData
 
-@Database(entities = [AirDataRecordEntity::class], version = 1)
+@Database(entities = [AirData::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
     abstract fun airDataRecordDao(): AirDataRecordDao
 }
