+++ app/src/main/java/tw/wesley/uiassignment/data/local/AirData.kt
@@ -1,11 +1,11 @@
-package tw.wesley.uiassignment.data.entities
+package tw.wesley.uiassignment.data.local
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "air_data_records")
-data class AirDataRecordEntity(
+data class AirData(
     @PrimaryKey @ColumnInfo(name = "site_id") val siteId: String,
     @ColumnInfo(name = "site_name") val siteName: String,
     @ColumnInfo(name = "county") val county: String,
