+++ app/src/main/java/tw/wesley/uiassignment/data/remote/AirDataResponse.kt
@@ -3,5 +3,5 @@ package tw.wesley.uiassignment.data.remote
 import com.squareup.moshi.Json
 
 data class AirDataResponse(
-    @Json(name = "records") val records: List<AirDataRecord>
+    @Json(name = "records") val records: List<RemoteAirDataRecord>
 )
