+++ app/src/main/java/tw/wesley/uiassignment/repo/AirDataRepository.kt
@@ -1,7 +1,7 @@
 package tw.wesley.uiassignment.repo
 
 import timber.log.Timber
-import tw.wesley.uiassignment.data.daos.AirDataRecordDao
+import tw.wesley.uiassignment.data.database.AirDataRecordDao
 import tw.wesley.uiassignment.data.mappers.AirDataRecordMapper
 import tw.wesley.uiassignment.network.AirQualityApiService
 import javax.inject.Inject
